refactor: remove unused fields from response models
