--- remote_datasource/src/main/java/com/baghdad/remoteDataSource/mapper/actor/ActorImagesResponseMapper.kt
@@ -1,13 +1,0 @@
-package com.baghdad.remoteDataSource.mapper.actor
-
-import com.baghdad.remoteDataSource.response.actor.ActorImagesResponse
-import com.baghdad.repository.model.actor.ActorImagesDto
-
-fun ActorImagesResponse.toDto(): ActorImagesDto {
-    return ActorImagesDto(
-        actorId = this.id?.toLong() ?: 0L,
-        images = this.profiles
-            ?.map { "https://image.tmdb.org/t/p/w500" + it.filePath }
-            ?: emptyList()
-    )
-}
