+++ remote_datasource/src/main/java/com/baghdad/remoteDataSource/mapper/actor/CastMemberDtoMapper.kt
@@ -2,7 +2,7 @@ package com.baghdad.remoteDataSource.mapper.actor
 
 import com.baghdad.remoteDataSource.response.CastMemberResponse
 import com.baghdad.repository.model.CastMemberDto
-import com.baghdad.repository.model.actor.ActorDto
+import com.baghdad.repository.model.ActorDto
 
 fun CastMemberResponse.toDto(): CastMemberDto {
     return CastMemberDto(
