+++ repository/src/main/java/com/baghdad/repository/datasource/remote/RemoteActorDataSource.kt
@@ -2,12 +2,11 @@ package com.baghdad.repository.datasource.remote
 
 import com.baghdad.repository.model.MovieDto
 import com.baghdad.repository.model.TvShowDto
-import com.baghdad.repository.model.actor.ActorDto
-import com.baghdad.repository.model.actor.ActorImagesDto
+import com.baghdad.repository.model.ActorDto
 
 interface RemoteActorDataSource {
     suspend fun getActorDetails(personId: Long): ActorDto
-    suspend fun getActorImages(personId: Long): ActorImagesDto
+    suspend fun getActorImages(personId: Long): List<String>
     suspend fun getActorMovies(personId: Long): List<MovieDto>
     suspend fun getActorTvShows(personId: Long): List<TvShowDto>
 }
