+++ repository/src/main/java/com/baghdad/repository/model/ActorDto.kt
@@ -1,4 +1,4 @@
-package com.baghdad.repository.model.actor
+package com.baghdad.repository.model
 
 data class ActorDto(
     val id : Long,
