+++ repository/src/main/java/com/baghdad/repository/model/CastMemberDto.kt
@@ -1,6 +1,6 @@
 package com.baghdad.repository.model
 
-import com.baghdad.repository.model.actor.ActorDto
+import com.baghdad.repository.model.ActorDto
 
 data class CastMemberDto(
     val actor: ActorDto,
