+++ repository/src/main/java/com/baghdad/repository/model/SearchResultDto.kt
@@ -1,6 +1,6 @@
 package com.baghdad.repository.model
 
-import com.baghdad.repository.model.actor.ActorDto
+import com.baghdad.repository.model.ActorDto
 
 data class SearchResultDto(
     val actors : List<ActorDto>,
