--- repository/src/main/java/com/baghdad/repository/model/actor/ActorImagesDto.kt
@@ -1,6 +1,0 @@
-package com.baghdad.repository.model.actor
-
-data class ActorImagesDto(
-    val actorId : Long,
-    val images : List<String>
-)
